requirements.txt
langchain
langchain_community
langchain-google-genai
python_dotenv
streamlit
langchainhub
yt_dlp
ffmpeg
aiodns
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
aleph-alpha-client
altair
annotated-types
anyio
async-timeout
attrs
backoff
bcrypt
blinker
boto3
botocore
Brotli
cachetools
certifi
cffi
charset-normalizer
click
colorama
coloredlogs
dataclasses-json
deeplake
dill
distro
exceptiongroup
faiss-cpu
ffmpeg
filelock
flatbuffers
frozenlist
fsspec
gitdb
GitPython
gpt-index
greenlet
grpcio
h11
httpcore
httptools
httpx
humanfriendly
humbug
idna
importlib_resources
Jinja2
jiter
jmespath
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
lz4
markdown-it-py
MarkupSafe
marshmallow
mdurl
monotonic
mpmath
multidict
multiprocess
mutagen
mypy-extensions
narwhals
numpy
onnxruntime
orjson
overrides
packaging
pandas
pathos
pillow
posthog
pox
ppft
propcache
protobuf
pulsar-client
pyarrow
pycares
pycparser
pycryptodomex
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
PyJWT
PyPDF2
PyPika
pyreadline3
python-dateutil
pytube
pytz
PyYAML
referencing
regex
requests
requests-toolbelt
rich
rpds-py
s3transfer
safetensors
shellingham
six
smmap
sniffio
SQLAlchemy
starlette
sympy
tenacity
tiktoken
tokenizers
toml
tornado
tqdm
transformers
typer
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
watchdog
watchfiles
websockets
yarl
yt-dlp